fix issue on WDL: wide category all set to missing category index, using double instead of float (validate model performance at fist, then future move to float)

+++ src/main/java/ml/shifu/shifu/core/dtrain/wdl/BinaryWDLSerializer.java
@@ -50,9 +50,9 @@ public class BinaryWDLSerializer {
             // version
             dos.writeInt(CommonConstants.WDL_FORMAT_VERSION);
             dos.writeUTF(modelConfig.getAlgorithm());
-            // Reserved two float field, one double field and one string field
-            dos.writeFloat(0.0f);
-            dos.writeFloat(0.0f);
+            // Reserved two double field, one double field and one string field
+            dos.writeDouble(0.0f);
+            dos.writeDouble(0.0f);
             dos.writeDouble(0.0d);
             dos.writeUTF("Reserved field");
 
@@ -72,9 +72,9 @@ public class BinaryWDLSerializer {
 
             // version
             fos.writeInt(CommonConstants.WDL_FORMAT_VERSION);
-            // Reserved two float field, one double field and one string field
-            fos.writeFloat(0.0f);
-            fos.writeFloat(0.0f);
+            // Reserved two double field, one double field and one string field
+            fos.writeDouble(0.0f);
+            fos.writeDouble(0.0f);
             fos.writeDouble(0.0d);
             fos.writeUTF("Reserved field");
